# Conflicts:
#	src/functionalTest/java/uk/gov/hmcts/reform/translate/TranslationServiceTestAutomationAdapter.java

+++ src/main/java/uk/gov/hmcts/reform/translate/repository/DictionaryRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 @Transactional(propagation = Propagation.REQUIRES_NEW)
 @Repository
-public interface DictionaryRepository extends CrudRepository<DictionaryEntity, Integer> {
+public interface DictionaryRepository extends CrudRepository<DictionaryEntity, Long> {
 
     Optional<DictionaryEntity> findByEnglishPhrase(String englishPhrase);
